+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * `Randomizer`
   * `getBytes($length)`
   * `shuffleArray($array)`
+  * `shuffleBytes($bytes)`
 * `Mt19937`
   * Serialization and unserialization are now compatible with PHP 8.2
     if performed in PHP 7.4+